Refactor affinity/client-mappings logic into separate subpackages.
