+++ clustering/ejb/extension/src/main/java/org/wildfly/extension/clustering/ejb/ClientMappingsRegistryProviderServiceConfigurator.java
@@ -29,7 +29,7 @@ import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceController;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceTarget;
-import org.wildfly.clustering.ejb.ClientMappingsRegistryProvider;
+import org.wildfly.clustering.ejb.remote.ClientMappingsRegistryProvider;
 import org.wildfly.clustering.service.FunctionalService;
 
 import java.util.function.Consumer;
