+++ clustering/ejb/infinispan/src/main/java/org/wildfly/clustering/ejb/infinispan/network/ClientMappingMarshaller.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.clustering.ejb.infinispan;
+package org.wildfly.clustering.ejb.infinispan.network;
 
 import java.io.IOException;
 import java.net.Inet4Address;
@@ -33,6 +33,7 @@ import org.wildfly.clustering.marshalling.protostream.ProtoStreamWriter;
 import org.wildfly.common.net.Inet;
 
 /**
+ * ProtoStream marshaller for a {@link ClientMapping}.
  * @author Paul Ferraro
  */
 public class ClientMappingMarshaller implements ProtoStreamMarshaller<ClientMapping> {
