+++ clustering/ejb/infinispan/src/main/java/org/wildfly/clustering/ejb/infinispan/network/ClientMappingsRegistryEntryMarshaller.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.ejb.infinispan;
+package org.wildfly.clustering.ejb.infinispan.network;
 
 import java.io.IOException;
 import java.util.LinkedList;
