+++ clustering/ejb/infinispan/src/test/java/org/wildfly/clustering/ejb/infinispan/network/ClientMappingMarshallerTestCase.java
@@ -20,13 +20,14 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.ejb.infinispan;
+package org.wildfly.clustering.ejb.infinispan.network;
 
 import java.net.InetAddress;
 
 import org.jboss.as.network.ClientMapping;
 import org.junit.Assert;
 import org.junit.Test;
+import org.wildfly.clustering.ejb.infinispan.network.ClientMappingMarshaller;
 import org.wildfly.clustering.marshalling.Tester;
 import org.wildfly.clustering.marshalling.protostream.ProtoStreamTesterFactory;
