+++ clustering/ejb/spi/src/main/java/org/wildfly/clustering/ejb/remote/RemoteEjbRequirement.java
@@ -19,21 +19,21 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.clustering.ejb;
+package org.wildfly.clustering.ejb.remote;
 
 import org.jboss.as.clustering.controller.RequirementServiceNameFactory;
 import org.jboss.as.clustering.controller.ServiceNameFactory;
 import org.jboss.as.clustering.controller.ServiceNameFactoryProvider;
 import org.wildfly.clustering.service.Requirement;
 
-public enum EjbRequirement implements Requirement, ServiceNameFactoryProvider {
+public enum RemoteEjbRequirement implements Requirement, ServiceNameFactoryProvider {
     CLIENT_MAPPINGS_REGISTRY_PROVIDER("org.wildfly.clustering.ejb.client-mappings-registry-provider", ClientMappingsRegistryProvider.class)
     ;
     private final String name;
     private final Class<?> type;
     private final ServiceNameFactory factory;
 
-    EjbRequirement(final String name, final Class<?> type) {
+    RemoteEjbRequirement(final String name, final Class<?> type) {
         this.name = name;
         this.type = type;
         this.factory = new RequirementServiceNameFactory(this);
